quiz_bot/pyproject.toml
[project]
name = "quiz-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=22.3",
]
